requirements.txt
PyYAML>=6.0
aiohttp>=3.8.1 